+++ pytket/extensions/quantinuum/backends/quantinuum.py
@@ -26,6 +26,7 @@ import re
 from typing import Dict, List, Set, Optional, Sequence, Union, Any, cast, Tuple
 from uuid import uuid1
 import warnings
+import datetime
 
 import numpy as np
 import requests
@@ -229,7 +230,7 @@ class QuantinuumBackendCompilationConfig:
 @cache
 def have_pecos() -> bool:
     try:
-        import pytket_pecos  # type: ignore
+        import pytket_pecos  # type: ignore # noqa # pylint: disable=unused-import
 
         return True
     except ImportError:
@@ -371,6 +372,7 @@ class QuantinuumBackend(Backend):
             dct1["system_type"] = "local_emulator"
             dct1.pop("emulator", None)
             dct1["batching"] = False
+        dct1["cl_reg_width"] = 32 if n_qubits <= 32 else 64
         return BackendInfo(
             name=cls.__name__,
             device_name=name + "LE" if local_emulator else name,
@@ -444,6 +446,95 @@ class QuantinuumBackend(Backend):
         api_handler._response_check(res, "get machine status")
         return str(res.json()["state"])
 
+    def get_calendar(
+        self,
+        start_date: datetime.datetime,
+        end_date: datetime.datetime,
+        localise: bool = True,
+    ) -> List[Dict[str, Any]]:
+        r"""Retrieves the Quantinuum H-Series operational calendar
+        for the period specified by start_date and end_date.
+        The calendar data returned is for the local timezone of the
+        end-user.
+
+        The output is a sorted list of dictionaries. Each dictionary is an
+        event on the operational calendar for the period specified by the
+        end-user. The output from this function can be readily used
+        to instantiate a pandas.DataFrame.
+
+        The dictionary has the following properties.
+        * 'start-date': The  start date and start time as a datetime.datetime object.
+        * 'end-date': The end date and end time as a datetime.datetime object.
+        * 'machine': A string specifying the H-Series device attached to the event.
+        * 'event-type': The type of event as a string. The value `online` denotes queued
+            access to the device, and the value `reservation` denotes priority access
+            for a particular organisation.
+        * 'organization': If the 'event-type' is assigned the value 'reservation', the
+            organization with reservation access is specified. Only users within an
+            organization have visibility on organization reservations. Otherwise,
+            organization is listed as 'fairshare', which means all users from all
+            organizations are able to submit jobs to the Fairshare queue during this
+            period.
+
+        :param start_date: The start date as datetime.date object
+            for the period to return the operational calendar.
+        :param end_date: The end date as datetime.date object
+            for the period to return the operational calendar.
+        :param localise: Apply localization to the datetime based
+            on the end-users time zone. Default is True. Disable by
+            setting False.
+        :return: A list of events from the H-Series operational calendar,
+            sorted by the `start-date` of each event. Each event is a python
+            dictionary.
+        :return_type: List[Dict[str, str]]
+        :raises: RuntimeError if an emulator or syntax-checker is specified
+        :raises: ValueError if the argument `start_date` or `end_date` are not
+            datetime.datetime objects.
+        """
+
+        if not isinstance(start_date, datetime.datetime) or not isinstance(
+            end_date, datetime.datetime
+        ):
+            raise ValueError(
+                "start_date and end_date must be datetime.datetime objects."
+            )
+
+        if self._device_name.endswith("E") | self._device_name.endswith("SC"):
+            raise RuntimeError(
+                f"Error requesting data for {self._device_name}. Emulators (E) \
+                and Syntax Checkers (SC) are online 24/7. Calendar \
+                information not available."
+            )
+
+        l4_calendar_data = self.api_handler.get_calendar(
+            start_date.date().isoformat(), end_date.date().isoformat()
+        )
+        calendar_data = []
+
+        for l4_event in l4_calendar_data:
+            device_name = l4_event["machine"]
+            if device_name != self._device_name:
+                continue
+            dt_start = _convert_datetime_string(
+                l4_event["start-date"]
+            )  # datetime in UTC tz
+            dt_end = _convert_datetime_string(
+                l4_event["end-date"]
+            )  # datetime in UTC tz
+            if localise:  # Apply timezone localisation on UTC datetime
+                dt_start = dt_start.astimezone()
+                dt_end = dt_end.astimezone()
+            event = {
+                "start-date": dt_start,
+                "end-date": dt_end,
+                "machine": device_name,
+                "event-type": l4_event["event-type"],
+                "organization": l4_event.get("organization", "fairshare"),
+            }
+            calendar_data.append(event)
+        calendar_data.sort(key=lambda item: item["start-date"])  # type: ignore
+        return calendar_data
+
     @property
     def backend_info(self) -> Optional[BackendInfo]:
         if self._backend_info is None and not self._MACHINE_DEBUG:
@@ -491,10 +582,7 @@ class QuantinuumBackend(Backend):
             return False
         info = self.backend_info
         assert info is not None
-        if info.get_misc("system_type") == "local_emulator":
-            return True
-        else:
-            return False
+        return bool(info.get_misc("system_type") == "local_emulator")
 
     def rebase_pass(self) -> BasePass:
         assert self.compilation_config.target_2qb_gate in self.two_qubit_gate_set
@@ -519,8 +607,6 @@ class QuantinuumBackend(Backend):
         squash = auto_squash_pass({OpType.PhasedX, OpType.Rz})
         target_2qb_gate = self.compilation_config.target_2qb_gate
         assert target_2qb_gate is not None
-        # use default (perfect fidelities) for supported gates
-        fidelities: Dict[str, Any] = {}
         if target_2qb_gate == OpType.TK2:
             decomposition_passes = []
         elif target_2qb_gate == OpType.ZZPhase:
@@ -880,7 +966,13 @@ class QuantinuumBackend(Backend):
             else:
                 results_selection = []
                 if language == Language.QASM:
-                    quantinuum_circ = circuit_to_qasm_str(c0, header="hqslib1")
+                    quantinuum_circ = circuit_to_qasm_str(
+                        c0,
+                        header="hqslib1",
+                        maxwidth=self.backend_info.misc["cl_reg_width"]
+                        if self.backend_info
+                        else 32,
+                    )
                     used_scratch_regs = _used_scratch_registers(quantinuum_circ)
                     for name, count in Counter(
                         bit.reg_name
@@ -897,6 +989,7 @@ class QuantinuumBackend(Backend):
                     for name, count in Counter(bit.reg_name for bit in c0.bits).items():
                         for i in range(count):
                             results_selection.append((name, i))
+
                     quantinuum_circ = b64encode(
                         cast(
                             bytes,
@@ -1146,7 +1239,8 @@ class QuantinuumBackend(Backend):
             if self.is_local_emulator:
                 if not have_pecos():
                     raise RuntimeError(
-                        "Local emulator not available: try installing with the `pecos` option."
+                        "Local emulator not available: \
+try installing with the `pecos` option."
                     )
                 from pytket_pecos import Emulator
 
@@ -1363,3 +1457,18 @@ def _parse_status(response: Dict) -> CircuitStatus:
     }
     message = json.dumps(msgdict)
     return CircuitStatus(_STATUS_MAP[h_status], message)
+
+
+def _convert_datetime_string(datetime_string: str) -> datetime.datetime:
+    year, month, day = list(map(int, datetime_string[:10].split("-")))
+    hour, minute, second = list(map(int, datetime_string[11:].split(":")))
+    dt = datetime.datetime(
+        year=year,
+        month=month,
+        day=day,
+        hour=hour,
+        minute=minute,
+        second=second,
+        tzinfo=datetime.timezone.utc,
+    )
+    return dt