+++ docs/changelog.rst
@@ -1,11 +1,16 @@
 Changelog
 ~~~~~~~~~
 
-Unreleased
-----------
+0.31.0 (March 2024)
+-------------------
 
-* Update pytket_pecos version requirement to 0.1.18.
+* Updated pytket version requirement to 1.26.
+* Add method to enable visibility of 
+Quantinuum H-Series operational calendar.
+* Update pytket_pecos version requirement to 0.1.19.
 * Support TK2 as native gate.
+* Update pytket version requirement to 1.26.
+* Update pytket-qir version requirement to 0.9.
 
 0.30.0 (February 2024)
 ----------------------